# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Платформа для удаленного управления системой контроля доступа шлагбаумов.",
    "description": "Спецификация API для системы контроля доступа шлагбаумов.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/api",
      "description": "Local docker prod server"
    },
    {
      "url": "http://localhost:8000/api",
      "description": "Local development server"
    },
    {
      "url": "http://46.8.225.185/api",
      "description": "Virtual dev docker server"
    },
    {
      "url": "https://gsm-barrier.ru/api",
      "description": "Production API"
    }
  ],
...
  "paths": {
...
    "/phones/{phone_id}/schedule/": {
      "get": {
        "tags": [
          "Phones"
        ],
        "summary": "Получение расписания номера.",
        "description": "Позволяет получить текущее расписание номера телефона.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "phone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 42
            },
            "description": "`ID` номера телефона."
          }
        ],
        "responses": {
          "200": {
            "description": "Текущее расписание номера.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные параметры запроса."
          },
          "401": {
            "description": "Пользователь не авторизован."
          },
          "403": {
            "description": "Доступ запрещен."
          },
          "404": {
            "description": "Номер не найден."
          },
          "500": {
            "description": "Ошибка сервера. Попробуйте позже."
          }
        }
      },
      "put": {
        "tags": [
          "Phones"
        ],
        "summary": "Изменение расписания номера.",
        "description": "Позволяет изменить расписание номера телефона, только для телефонов которые работают по расписанию.\n\nМеняются только те значения, которые будут переданы в запросе.\n\nЕсли передать `null`, то расписание на этот день будет удалено.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "phone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 42
            },
            "description": "`ID` номера телефона."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "schedule": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Расписание изменено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные параметры запроса."
          },
          "401": {
            "description": "Пользователь не авторизован."
          },
          "403": {
            "description": "Выбранные изменения невозможно применить."
          },
          "404": {
            "description": "Номер не найден."
          },
          "500": {
            "description": "Ошибка сервера. Попробуйте позже."
          }
        }
      }
    },
    "/admin/phones/{phone_id}/schedule/": {
      "get": {
        "tags": [
          "Phones"
        ],
        "summary": "Получение расписания номера.",
        "description": "Позволяет администратору получить текущее расписание номера телефона.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "phone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 42
            },
            "description": "`ID` номера телефона."
          }
        ],
        "responses": {
          "200": {
            "description": "Текущее расписание номера.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные параметры запроса."
          },
          "401": {
            "description": "Пользователь не авторизован."
          },
          "403": {
            "description": "Доступ запрещен."
          },
          "404": {
            "description": "Номер не найден."
          },
          "500": {
            "description": "Ошибка сервера. Попробуйте позже."
          }
        }
      },
      "put": {
        "tags": [
          "Phones"
        ],
        "summary": "Изменение расписания номера.",
        "description": "Позволяет администратору изменить расписание номера телефона, только для телефонов которые работают по расписанию.\n\nМеняются только те значения, которые будут переданы в запросе.\n\nЕсли передать `null`, то расписание на этот день будет удалено.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "phone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 42
            },
            "description": "`ID` номера телефона."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "schedule": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Расписание изменено.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные параметры запроса."
          },
          "401": {
            "description": "Пользователь не авторизован."
          },
          "403": {
            "description": "Выбранные изменения невозможно применить."
          },
          "404": {
            "description": "Номер не найден."
          },
          "500": {
            "description": "Ошибка сервера. Попробуйте позже."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ScheduleResponse": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "integer",
            "description": "`ID` номера телефона.",
            "example": 42
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          }
        }
      },
      "Schedule": {
        "type": "object",
        "description": "Расписание работы номера.",
        "properties": {
          "monday": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleInterval"
            }
          },
          "tuesday": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleInterval"
            }
          },
          "wednesday": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleInterval"
            }
          },
          "thursday": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleInterval"
            }
          },
          "friday": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleInterval"
            }
          },
          "saturday": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleInterval"
            }
          },
          "sunday": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleInterval"
            }
          }
        }
      },
      "ScheduleInterval": {
        "type": "object",
        "properties": {
          "start_time": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "description": "Время начала в формате HH:MM.",
            "example": "08:00"
          },
          "end_time": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "description": "Время окончания в формате HH:MM.",
            "example": "12:00"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}